requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-filter==2.2.0
djangorestframework==3.10.2
idna==2.8
Markdown==3.1.1
pytz==2019.1
requests==2.5.0
six==1.8.0
sqlparse==0.3.0
swapi==0.1.3
ujson==1.33
urllib3==1.25.3
